requirements.txt
accumulation_tree==0.6.2
asgiref==3.8.1
boto3==1.21.32
botocore==1.24.46
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.9.0
django-extensions==3.2.1
django-rest-framework==0.1.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-writable-nested==0.7.0
idna==3.7
jmespath==1.0.1
jsonfield==3.1.0
mysqlclient==2.1.1
numpy==1.26.4
pandas==2.2.2
pybrake==1.10.1
pyfcm==1.5.4
PyJWT==2.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyudorandom==1.0.0
requests==2.31.0
s3transfer==0.5.2
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
tdigest==0.5.2.2
tzdata==2024.1
urllib3==1.26.18
whitenoise==6.2.0
